Add users_cron_allow variable to manage cron permissions
If users are deployed with this role, the file /etc/cron.allow is
created, even when `cron_allow` is unset. But if /etc/cron.allow exists,
only the users listed in that file are allowed to run cron. In other
word, if an empty `/etc/cron.allow` is created no user is allow to run
cron.

To disable the deployment of this file the variable users_cron_allow
should be set to 'false'.

Closes: robertdebock/ansible-role-users#5